help: Implementation of a help mailbox
Help system means to parse Markdown files in a directory (help mailbox),
show valid results threaded by chapter/section in the index view as help
documents and view those "messages" in pager by pressing <Enter> on it.

The help mailbox (helpbox) works local, e.g. with a clone of [1][2], for
testing, is read-only and all files are handled accordingly.

This commit implements:
  - a new DT_PATH option $help_doc_dir, specifies a folder that contains
    the Markdown help files to process, defaults to (PKGDOCDIR "/help")
  - a new function <help-box> (OP_HELP_BOX), bound to "<Esc>H" or may be
    invoked directly by ":exec help-box" or by "<change-folder> help://"
  - a rudimentary "jump to chapter/section/file" functionality from the
    given (and first matching) help URI, e.g. "help://sidebar/intro"
  - hard-coded option for caching a list of valid help documents between
    help invocation; whether or not to link chapter threads upwards; how
    many lines of a YAML file header to read

References:
  - [1] <https://github.com/neomutt/test-doc>
  - [2] <https://github.com/neomutt/neomutt.github.io>

Co-authored-by: Floyd Anderson <f.a@31c0.net>
Co-authored-by: Tran Manh Tu <xxlaguna93@gmail.com>